+++ ledger/ledger-api-integration-tests/src/test/itsuite/scala/com/digitalasset/platform/tests/integration/ledger/api/commands/CommandTransactionChecksNo1866IT.scala
@@ -4,21 +4,16 @@
 package com.digitalasset.platform.tests.integration.ledger.api.commands
 
 import java.io.File
+
 import com.digitalasset.daml.bazeltools.BazelRunfiles.rlocation
-import com.digitalasset.daml.lf.language.{
-  LanguageVersion,
-  LanguageMajorVersion => LMajV,
-  LanguageMinorVersion => LMinV
-}
-import com.digitalasset.platform.apitesting.CommandTransactionChecksHighLevel
+import com.digitalasset.daml.lf.language.{LanguageVersion, LanguageMajorVersion => LMajV, LanguageMinorVersion => LMinV}
+import com.digitalasset.platform.tests.integration.ledger.api.ContractKeysIT
 
 /** Runs tests with an older version of the language that does not include changes
   * performed in #1866 -- specifically the fact that the submitter must be in
   * lookup maintainers.
-  *
-  * TODO extract version-dependent parts out of the huge `CommandTransactionChecks`.
   */
-class CommandTransactionChecksNo1866IT extends CommandTransactionChecksHighLevel {
+class CommandTransactionChecksNo1866IT extends ContractKeysIT {
   override protected val languageVersion: LanguageVersion =
     LanguageVersion(LMajV.V1, LMinV.fromProtoIdentifier("5"))
   override protected lazy val config: Config = Config.default.copy(
